fix a broken FAT test.
